+++ tests/test_lint.py
@@ -176,6 +176,22 @@ def test_baseline_restriction():
     assert set(["BASELINE_PING"]) == set(v.check_name for v in nits)
 
 
+def test_default_warning():
+    contents = [
+        {"user_data": {"counter": {"type": "counter", "send_in_pings": ["default"]}}}
+    ]
+    contents = [util.add_required(x) for x in contents]
+    all_metrics = parser.parse_objects(contents)
+
+    errs = list(all_metrics)
+    assert len(errs) == 0
+
+    nits = lint.lint_metrics(all_metrics.value, {"output_format": "javascript"})
+
+    assert len(nits) == 1
+    assert set(["DEFAULT_PING"]) == set(v.check_name for v in nits)
+
+
 def test_misspelling_pings():
     contents = [
         {